+++ doc/release_notes.md
@@ -18,6 +18,7 @@
 - The new "Watch Log File" mode can be used to edit log files and have them update their view and scroll to the end, when new log entries
   are added (seee Linux tail -f). In this mode editing the file is not possible - it is read-only in this case to avoid concurrent external
   and internal changes.
+- One can now move the current selection (or - if no selection exists - current line) up and down using Alt+UP and Alt+DOWN respectively.
 
 ### Improved
   The temp path of PKS Edit may be configured explicitly to point to an arbitrary directory, but will also default now to a directory named